+++ Source/Graphics/Mesh.h
@@ -10,13 +10,13 @@ public:
     const std::vector<float>& getVertices() const { return vertices; }
     uint32_t getVerticesSize() const { return sizeof(vertices[0]) * vertices.size(); }
 
-    void setIndices(const std::vector<uint32_t> indices);
-    const std::vector<uint32_t>& getIndices() const { return indices; }
+    void setIndices(const std::vector<uint16_t> indices);
+    const std::vector<uint16_t>& getIndices() const { return indices; }
     uint32_t getIndicesSize() const { return sizeof(indices[0]) * indices.size(); }
 
     void draw();
 
 private:
     std::vector<float> vertices;
-    std::vector<uint32_t> indices;
+    std::vector<uint16_t> indices;
 };
